app/requirements.txt
# Copyright 2025 Google LLC.
#
# Licensed under the Apache License, Version 2.0 (the "License");
# you may not use this file except in compliance with the License.
# You may obtain a copy of the License at
#
#     https://www.apache.org/licenses/LICENSE-2.0
#
# Unless required by applicable law or agreed to in writing, software
# distributed under the License is distributed on an "AS IS" BASIS,
# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
# See the License for the specific language governing permissions and
# limitations under the License.

mesop==0.14.1
uvicorn==0.34.0
fastapi==0.115.6
google-auth==2.37.0
requests==2.32.3
google-cloud-logging==3.11.3
absl-py==2.1.0
google-api-core==2.24.0
aiohttp==3.11.10
pillow==11.1.0